requirements.txt
anyio==4.6.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastcore==1.7.11
fastlite==0.0.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.1
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.6.9
python-multipart==0.0.12
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlite-minutils==3.37.0.post3
starlette==0.39.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
